scripts/requirements.txt
# Python dependencies for CV processing scripts

# Supabase client
supabase>=2.0.0

# HTTP client for downloading files
httpx>=0.24.0

# PDF parsing
PyMuPDF>=1.23.0

# Environment variables
python-dotenv>=1.0.0
